requirements.txt
numpy==1.17.2
torchvision==0.4.2
torch==1.3.1
animeface==1.1.0
Pillow==7.0.0
